backend-01/requirements/base.txt
Django==3.2.21
wagtail==4.2.3
django-environ==0.10.0
djangorestframework==3.12.0
django-cors-headers==3.14.0
django-filter==2.4.0
django-autoslug==1.9.9
django-countries==7.5.1
django-mptt==0.14.0
django-phonenumber-field==7.0.2
phonenumbers==8.13.10
drf-yasg==1.21.5
wagtail-factories==4.1.0
wagtail-headless-preview==0.6.0

Pillow==9.5.0

argon2-cffi==21.3.0

pytz==2023.3

redis==4.5.4
celery==5.2.7
flower==1.2.0
django-celery-email==3.0.0

djoser==2.2.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0

django-haystack==3.2.1
Whoosh==2.7.4
drf-haystack==1.5.6


psycopg2-binary== 2.9.3

factory-boy==3.3.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
Faker==18.3.1
pytest-cov==4.0.0